\name{BCexdata}
\alias{BCexdata}
\alias{B1}
\alias{B2}
\title{
 BC population dataset    
}
\description{
 The phenotype of BC population . 
}
\usage{
  data(BCexdata)
}
\details{
 Dataset input of BCFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
