\name{BILexdata}
\alias{BILexdata}
\alias{BIL}
\title{
 BIL population dataset    
}
\description{
 The phenotype of BIL population . 
}
\usage{
  data(BILexdata)
}
\details{
 Dataset input of BILFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
