% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{check_antimerid}
\alias{check_antimerid}
\title{Check to se if crusie track crosses Anti-meridion}
\usage{
check_antimerid(df)
}
\arguments{
\item{df}{tibble from and read_ source}
}
\description{
Check to se if crusie track crosses Anti-meridion
}
