% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_var}
\alias{check_var}
\title{Check if variable exists in data frame}
\usage{
check_var(df, var)
}
\arguments{
\item{var}{}
}
\description{
Assigns the variable to "val" if found, otherwise stops with error
}
