% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\name{cnv_to_csv}
\alias{cnv_to_csv}
\title{Convert a cnv file to a csv file}
\usage{
cnv_to_csv(in_file, out_fold, overwrite = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Convert a cnv file to a csv file
}
