% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\name{cnv_to_csv_fold}
\alias{cnv_to_csv_fold}
\title{Convert all cnv files in a folder to csv}
\usage{
cnv_to_csv_fold(in_fold, out_fold, ...)
}
\arguments{
\item{...}{}
}
\description{
Convert all cnv files in a folder to csv
}
