\name{DHexdata}
\alias{DHexdata}
\alias{DH}
\title{
 DH population dataset    
}
\description{
 The phenotype of DH population . 
}
\usage{
  data(DHexdata)
}
\details{
 Dataset input of DHFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
