% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{extract_bathy}
\alias{extract_bathy}
\title{Extract ETOPO5 bathymetry for cruise region}
\usage{
extract_bathy(df, lonran = NULL, latran = NULL)
}
\arguments{
\item{latran}{}
}
\description{
Extract ETOPO5 bathymetry for cruise region
}
