\name{F23exdata}
\alias{F23exdata}
\alias{F23}
\title{
 F23 population dataset    
}
\description{
 The phenotype of F23 population . 
}
\usage{
  data(F23exdata)
}
\details{
 Dataset input of F23Fun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
