\name{F2exdata}
\alias{F2exdata}
\alias{F2}
\title{
 F2 population dataset    
}
\description{
 The phenotype of F2 population . 
}
\usage{
  data(F2exdata)
}
\details{
 Dataset input of F2Fun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
