% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\name{find_datasheet}
\alias{find_datasheet}
\title{Find the name of a datasheet}
\usage{
find_datasheet(files, pattern)
}
\arguments{
\item{pattern}{}
}
\description{
Takes in a list of files and a pattern and finds the best fit
}
