\name{G3DHexdata}
\alias{G3DHexdata}
\alias{G3DH-P1}
\alias{G3DH-P2}
\alias{G3DH-DH}
\title{
 G3DH population dataset    
}
\description{
 The phenotype of G3DH population . 
}
\usage{
  data(G3DHexdata)
}
\details{
 Dataset input of G3DHFun function.
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
