\name{G4F2exdata}
\alias{G4F2exdata}
\alias{G4F2-P1}
\alias{G4F2-P2}
\alias{G4F2-F1}
\alias{G4F2-F2}
\title{
 G4F2 population dataset    
}
\description{
 The phenotype of G4F2 population . 
}
\usage{
  data(G4F2exdata)
}
\details{
 Dataset input of G4F2Fun function.
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
