\name{G5BCexdata}
\alias{G5BCexdata}
\alias{G5BC-P1}
\alias{G5BC-P2}
\alias{G5BC-F1}
\alias{G5BC-B1}
\alias{G5BC-B2}
\title{
 G5BC population dataset    
}
\description{
 The phenotype of G5BC population . 
}
\usage{
  data(G5BCexdata)
}
\details{
 Dataset input of G5BCFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
