\name{G5BCFexdata}
\alias{G5BCFexdata}
\alias{G5BCF-P1}
\alias{G5BCF-P2}
\alias{G5BCF-F1}
\alias{G5BCF-B12}
\alias{G5BCF-B22}
\title{
 G5BCF population dataset    
}
\description{
 The phenotype of G5BCF population . 
}
\usage{
  data(G5BCFexdata)
}
\details{
 Dataset input of G5BCFFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
