\name{G5exdata}
\alias{G5exdata}
\alias{G5-P1}
\alias{G5-P2}
\alias{G5-F1}
\alias{G5-F2}
\alias{G5-F23}
\title{
 G5 population dataset    
}
\description{
 The phenotype of G5 population . 
}
\usage{
  data(G5exdata)
}
\details{
 Dataset input of G5Fun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
