\name{G6exdata}
\alias{G6exdata}
\alias{G6-P1}
\alias{G6-P2}
\alias{G6-F1}
\alias{G6-F2}
\alias{G6-B1}
\alias{G6-B2}
\title{
 G6 population dataset    
}
\description{
 The phenotype of G6 population . 
}
\usage{
  data(G6exdata)
}
\details{
 Dataset input of G6Fun function.
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
