\name{G6Fexdata}
\alias{G6Fexdata}
\alias{G6F-P1}
\alias{G6F-P2}
\alias{G6F-F1}
\alias{G6F-B12}
\alias{G6F-B22}
\alias{G6F-F23}
\title{
 G6F population dataset    
}
\description{
 The phenotype of G6F population . 
}
\usage{
  data(G6Fexdata)
}
\details{
 Dataset input of G6FFun function.
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
