% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{generate_swath}
\alias{generate_swath}
\title{Generate swath for plotting sections}
\usage{
generate_swath(lon, lat, width)
}
\arguments{
\item{width}{}
}
\description{
Generate swath for plotting sections
}
