% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_ctd_meta}
\alias{get_ctd_meta}
\title{Get a vector of a paricular metadata variable from a list of oce ctd objects}
\usage{
get_ctd_meta(ctd, meta = "station")
}
\arguments{
\item{var}{}
}
\description{
Get a vector of a paricular metadata variable from a list of oce ctd objects
}
