% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_ctd_var}
\alias{get_ctd_var}
\title{Get a matrix of a paricular variable from a list of oce ctd objects}
\usage{
get_ctd_var(ctd, var = "temperature")
}
\arguments{
\item{var}{}
}
\description{
Get a matrix of a paricular variable from a list of oce ctd objects
}
