% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{grid_section}
\alias{grid_section}
\title{Grid section data onto a regular dist-depth grid}
\usage{
grid_section(df, xo = NULL, zo = NULL, ...)
}
\arguments{
\item{df}{a data frame with three columns for var, dist, depth}

\item{xo}{distance vector}

\item{zo}{depth vector}
}
\description{
Grid section data onto a regular dist-depth grid
}
