% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{grid_section_hydro}
\alias{grid_section_hydro}
\title{Grid a bottle derived value}
\usage{
grid_section_hydro(df, var = "chla", select = NULL, xo = NULL,
  yo = NULL, along_path = T)
}
\arguments{
\item{df}{data frame read in using read_hydrocast}

\item{var}{}
}
\description{
Grid a bottle derived value
}
