% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adcp.R
\name{interp_adcp}
\alias{interp_adcp}
\title{Interpolate an adcp field to different depths}
\usage{
interp_adcp(adcp, dvec, vars = c("u", "v", "backscat", "quality",
  "percent"))
}
\arguments{
\item{vars}{}
}
\description{
Interpolate an adcp field to different depths
}
