% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{is_sea_struct}
\alias{is_sea_struct}
\title{Check for existence of standard SEA dataframe organization}
\usage{
is_sea_struct(df)
}
\arguments{
\item{df}{}
}
\description{
Check for existence of standard SEA dataframe organization
}
