% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_tools.R
\name{kml_to_delim}
\alias{kml_to_delim}
\title{Convert kml station locations to output coordinates}
\usage{
kml_to_delim(kml_file, file_out = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Convert kml station locations to output coordinates
}
