% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{make_base_map}
\alias{make_base_map}
\title{Make a base map object}
\usage{
make_base_map(df = NULL, lonlim = NULL, latlim = NULL, plot_bathy = F,
  high_res = F)
}
\arguments{
\item{latlim}{}
}
\description{
Make a base map object
}
