% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_layers.R
\name{make_points}
\alias{make_points}
\title{Generic generator of geom_points from data}
\usage{
make_points(df, var = "temp", data_source = "elg", step = 1, size = 2,
  ran_val = NULL, ran_qua = c(0.01, 0.99))
}
\arguments{
\item{colormap}{}
}
\description{
Generic generator of geom_points from data
}
