% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ctd.R
\name{make_section}
\alias{make_section}
\title{Make a section from a list of CTD objects}
\usage{
make_section(ctd, select = 1:length(ctd))
}
\arguments{
\item{ctd}{list of ctd objects as generated by read_ctd_fold}

\item{select}{selected stations within the list to use in section}
}
\description{
Make a section from a list of CTD objects
}
