% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_layers.R
\name{make_subplots}
\alias{make_subplots}
\title{Creates subplots from list of ggplot objects}
\usage{
make_subplots(obj, nrow = 1, ncol = 1)
}
\arguments{
\item{ncol}{}
}
\description{
Creates subplots from list of ggplot objects
}
