% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_layers.R
\name{make_track}
\alias{make_track}
\title{Make a cruise track ggplot geom}
\usage{
make_track(df, data_source = "elg")
}
\arguments{
\item{data_source}{}
}
\description{
Make a cruise track ggplot geom
}
