% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{make_vector_lonlat}
\alias{make_vector_lonlat}
\title{Turns vectors into lon and lat moved}
\usage{
make_vector_lonlat(lon, lat, u, v)
}
\arguments{
\item{v}{}
}
\description{
Turns vectors into lon and lat moved
}
