% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_layers.R
\name{make_vectors}
\alias{make_vectors}
\title{A geom of vector lines}
\usage{
make_vectors(df, data_source = "elg", field = "wind", step = 60,
  scale = 1)
}
\arguments{
\item{step}{}
}
\description{
A geom of vector lines
}
