% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAmisc.R
\name{makeFolders}
\alias{makeFolders}
\title{A function that creates the output folders in your home directory}
\usage{
makeFolders(cruiseID, isReport = F)
}
\arguments{
\item{cruiseID}{:string containing the name of the SEA cruise}

\item{isReport}{:T/F argument (default F) to determine if the outputs are for a cruise report}
}
\description{
Only really applies to author's account
}
\examples{
makeFolders()
}
