% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tension.R
\name{max_tension}
\alias{max_tension}
\title{Find maximum tension for wire files in a folder}
\usage{
max_tension(data_folder, cruiseID = NULL, width = 8, thresh = 100,
  plotFL = TRUE, check = TRUE)
}
\arguments{
\item{data_folder}{the folder where the tension files are kept}
}
\description{
Find maximum tension for wire files in a folder
}
