% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_ctd.R
\name{opll}
\alias{opll}
\title{Calculate 1% light level from CTD PAR data}
\usage{
opll(CTD)
}
\arguments{
\item{CTD}{CTD object created by read.ctd from the oce package}
}
\description{
Function reads in a CTD object and returns a
modelled 1% light level based on a fitted natural log.
}
\examples{
opll()

}
