% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\name{package_data}
\alias{package_data}
\title{Creates standard data structures of SEA data}
\usage{
package_data(root_folder, cruiseID = NULL, ...)
}
\arguments{
\item{root_folder}{}
}
\description{
Takes data in standard SEA archive format and extracts the correct files
and creates standardized data format
}
