% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_bubble}
\alias{plot_bubble}
\title{Make a bubble plot}
\usage{
plot_bubble(df, var, ran_val = NULL, ran_qua = c(0, 1), base_map = NULL,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Make a bubble plot
}
