% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_cruise_track_gg}
\alias{plot_cruise_track_gg}
\title{Plot a cruise track from an sea data frame containing lon and lat data}
\usage{
plot_cruise_track_gg(df, type = "elg", bathy = T, stations = NULL,
  reg = "", coastline = "fine", legend = T)
}
\arguments{
\item{df}{}
}
\description{
Plot a cruise track from an sea data frame containing lon and lat data
}
