% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_current}
\alias{plot_current}
\title{Plot surface from an ELG or Hourly file}
\usage{
plot_current(df, step = 1, scale = 0.005, ...)
}
\arguments{
\item{df}{}
}
\description{
Plot surface from an ELG or Hourly file
}
