% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_flowthrough}
\alias{plot_flowthrough}
\title{Plot the flow-through data on a leaflet map}
\usage{
plot_flowthrough(df, type = "temp", step = 60)
}
\arguments{
\item{type}{temp, sal or fluor currently}

\item{step}{gap between consecutive colored dots}

\item{dfdata}{frame read from read_elg or read_hourly}
}
\description{
Plot the flow-through data on a leaflet map
}
