% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_flowthru}
\alias{plot_flowthru}
\title{Plot flow through data}
\usage{
plot_flowthru(df, var = "temp", data_source = "elg", step = NULL,
  colormap = oce::oce.colorsTemperature(), ran_val = NULL,
  ran_qua = c(0.01, 0.99), title = stringr::str_to_title(var), ...)
}
\arguments{
\item{...}{}
}
\description{
Plot flow through data
}
