% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_neuston}
\alias{plot_neuston}
\title{Plot a neuston tow dataset on a map}
\usage{
plot_neuston(df, var = "biomass", data_source = "neuston", step = 1,
  colormap = oce::oce.colorsTemperature(), ran_val = NULL,
  ran_qua = c(0, 1), title = stringr::str_to_title(var),
  type = "bubble", ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a neuston tow dataset on a map
}
