% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_surfvals}
\alias{plot_surfvals}
\title{Plot flow through data}
\usage{
plot_surfvals(df, var = "temp", ran_val = NULL, ran_qua = c(0.01, 0.99),
  step = 60, base_map = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot flow through data
}
