% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plothydro}
\alias{plothydro}
\title{Plot Bottle data from Hydrocasts}
\usage{
plothydro(df, CTDs = NULL, plotmap = T, which = NULL)
}
\arguments{
\item{df}{data frome created using readSEAxls() on <cruiseID>_hyrdoworl.xls}

\item{CTDs}{location of CTD files for comparisson (not needed)}

\item{plotmap}{logical to indicate whether to plot a map}

\item{which}{list of parameters to plot}
}
\description{
Currently plots Nitrate, Phosphate and Chl-a data
}
\details{
Notes: which parameters:

1 - Chl-a

2 - Nitrate

3 - Phosphate
}
\examples{
plothydro()
}
