% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotmap3sec}
\alias{plotmap3sec}
\title{Plot CTD sections as map and three panels of T, S and rho}
\usage{
plotmap3sec(CTDs, ylim = c(0, 300), Tran = NULL, Sran = NULL,
  Dran = NULL, dist_vec = NULL)
}
\arguments{
\item{CTDs}{object containing CTD objects created using read.ctd from oce package.}

\item{ylim}{range of values for y axis.}

\item{Tran}{Range of Temperature}

\item{Sran}{Range of Salinity}

\item{Dran}{Range of Potential Density}
}
\description{
Plot CTD sections as map and three panels of T, S and rho
}
\examples{
plotmap3sec()
}
