% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotmapTS}
\alias{plotmapTS}
\title{Plot CTD sections as map and two panels of T and S}
\usage{
plotmapTS(CTDs, ylim = c(0, 1000))
}
\arguments{
\item{CTDs}{object containing CTD objects created using read.ctd from oce package.}

\item{ylim}{range of values for y axis.}
}
\description{
Plot CTD sections as map and two panels of T and S
}
\examples{
plotmapTS()
}
