% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotO2flsec}
\alias{plotO2flsec}
\title{Plot CTD auxilary data sections (fluo/oxygen) as two panels}
\usage{
plotO2flsec(CTDs)
}
\arguments{
\item{CTDs}{object containing CTD objects created using read.ctd from oce package.}
}
\description{
Plot CTD auxilary data sections (fluo/oxygen) as two panels
}
\examples{
plot02flsec()
}
