% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotRBRtow}
\alias{plotRBRtow}
\title{Plot section from RBR Tow-Yo data}
\usage{
plotRBRtow(df)
}
\arguments{
\item{df}{data frame create using readRBRtow()}
}
\description{
Plot section from RBR Tow-Yo data
}
\examples{
plotRBRtow()
}
