% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAbio}
\alias{plotSEAbio}
\title{Plot biomass distribution from neustron tows data}
\usage{
plotSEAbio(df, legloc = "topleft")
}
\arguments{
\item{df}{data frame created by using readSEAxls() on <cruiseID_Neuston.xls>.}

\item{legloc}{location of legend (passed to legend()).}
}
\description{
Plot biomass distribution from neustron tows data
}
\examples{
plotSEAbio()
}
