% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAbiod}
\alias{plotSEAbiod}
\title{Plot biodiveristy distribution from neustron tows data}
\usage{
plotSEAbiod(df, legloc = "topleft", mapPad = 0.1)
}
\arguments{
\item{df}{data frame created by using readSEAxls() on <cruiseID_Neuston.xls>.}

\item{legloc}{location of legend (passed to legend()).}
}
\description{
Plot biodiveristy distribution from neustron tows data
}
\examples{
plotSEAbiod()
}
