% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAbiodn}
\alias{plotSEAbiodn}
\title{Plot biomass boxplot between day and night from neustron tows data}
\usage{
plotSEAbiodn(df)
}
\arguments{
\item{df}{data frame created by using readSEAxls() on <cruiseID_Neuston.xls>.}
}
\description{
Plot biomass boxplot between day and night from neustron tows data
}
\examples{
plotSEAbiodn()
}
