% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAct}
\alias{plotSEAct}
\title{Plots map of cruise track}
\usage{
plotSEAct(df, type = "elg", bathy = T, stations = NULL, reg = "",
  coastline = "fine", legend = T)
}
\arguments{
\item{df}{data frame generated either by readSEAxls or readSEAelgthat contains GPS data}

\item{type}{choose what type of data is in df (can be "elg" or "hourly")}

\item{bathy}{logical for including background bathymetry (FIX: CURRENTLY STORED LOCALLY)}

\item{stations}{option to include the output from readbioll() or readCTDsll() for locations of stations}

\item{reg}{option to define a plotting region (calls reg2latlon).}
}
\description{
Plots map of cruise track
}
\examples{
plotSEAct()
}
