% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAcurr}
\alias{plotSEAcurr}
\title{Plot map of currents from hull-mounted ADCP data}
\usage{
plotSEAcurr(X, stp = 6, scale = 0.2, plotKEY = T, reg = "")
}
\arguments{
\item{X}{object created by readSEAadcp() or readSEAadcp_all().}

\item{stp}{integer indicating the number of timesteps to skip between subsequent arrows.}

\item{scale}{double to scale the vector arrows drawn}

\item{plotKEY}{logical plot or don't plot}

\item{reg}{option to define a plotting region (calls reg2latlon).}
}
\description{
Plot map of currents from hull-mounted ADCP data
}
\examples{
plotSEAcurr()
}
