% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAplastic}
\alias{plotSEAplastic}
\title{Plot plastic distribution from neustron tows data}
\usage{
plotSEAplastic(df, legloc = "topleft")
}
\arguments{
\item{df}{data frame created by using readSEAxls() on <cruiseID_Neuston.xls>.}

\item{legloc}{location of legend (passed to legend()).}
}
\description{
Plot plastic distribution from neustron tows data
}
\examples{
plotSEAplastic()
}
