% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{plotSEAsurf}
\alias{plotSEAsurf}
\title{Plot Surface Station data}
\usage{
plotSEAsurf(df, vars = c(1, 2, 3, 4), reg = "")
}
\arguments{
\item{df}{Data from created by using readSEAxls().}

\item{vars}{list of numbers prescribing which fields to plot (see notes).}

\item{reg}{internally prescribed regions that can be plotted for local maps (see reg2latlon).}
}
\description{
Plots data recorded in surfsamp.xls(x/m) once read in to a data frame using readSEAxls
}
\details{
Plots are made in subplots determined by the number of variables in vars

vars can have following properties:

1 - Nitrates

2 - Phosphates

3 - Extracted Chla

4 - pH

5 - Alkalinity

6 - Oxygen

Regs can currently have values prescribed in reg2latlon().
}
\examples{
plotSEAsurf()
}
