% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{prep_section_adcp}
\alias{prep_section_adcp}
\title{Prepare ADCP data for plotting section}
\usage{
prep_section_adcp(X, select = NULL, dist_vec = NULL, along_section = F,
  sec_lon = NULL, sec_lat = NULL, dx = 5, dz = 25, width = 10)
}
\arguments{
\item{sec_lon}{two point vector c(starting lon, ending lon)}

\item{sec_lat}{two point vector c(starting lat, ending lat)}

\item{dx}{distance between interpolated points along section in km}

\item{width}{width around which to gather data in km}
}
\description{
Prepare ADCP data for plotting section
}
