% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{prep_section_ctd}
\alias{prep_section_ctd}
\title{Prepare ctd objects to plot as a section}
\usage{
prep_section_ctd(sec, var = "temperature", select = NULL, dist_vec = NULL,
  along_section = F, sec_lon = NULL, sec_lat = NULL, dx = 5, dz = 5,
  width = 10)
}
\arguments{
\item{sec}{ctd list or section object to make into section}

\item{var}{variable to plot}
}
\description{
Prepare ctd objects to plot as a section
}
