% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{printTable}
\alias{printTable}
\title{Does the final table printing and saving after data has been edited}
\usage{
printTable(dfo, caption = "", SigF = NULL, secondRow = NULL,
  thirdRow = NULL, fileout = "~/Desktop/tableOutput.tex")
}
\arguments{
\item{dfo}{data frame of data ready to be exported}

\item{caption}{caption of table.}

\item{SigF}{vector of significant figures for each column}

\item{secondRow}{List of entries for second row of table headings if needed}

\item{thirdRow}{List of entries for third row of table headings if needed}

\item{fileout}{name and location of the file to be exported}
}
\description{
Should only be called from one of the SEA table functions
}
\examples{
printTable()
}
