% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adcp.R
\name{read_adcp_ens}
\alias{read_adcp_ens}
\title{Read ADCP file that is a winADCP output}
\usage{
read_adcp_ens(adcp_file)
}
\arguments{
\item{adcp_file}{}
}
\description{
Read ADCP file that is a winADCP output
}
