% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adcp.R
\name{read_adcp_ens_fold}
\alias{read_adcp_ens_fold}
\title{Read all ADCP Ensemble files in a folder and combine}
\usage{
read_adcp_ens_fold(adcp_fold)
}
\arguments{
\item{adcp_fold}{}
}
\description{
Read all ADCP Ensemble files in a folder and combine
}
