% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adcp.R
\name{read_adcp_fold}
\alias{read_adcp_fold}
\title{Read adcp files from a folder}
\usage{
read_adcp_fold(adcp_fold, pattern = "1-30")
}
\arguments{
\item{pattern}{}
}
\description{
Read adcp files from a folder
}
