% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ctd.R
\name{read_ctd}
\alias{read_ctd}
\title{Read and process CTD data}
\usage{
read_ctd(cnv_file, pmin = 5, p = 1, ...)
}
\arguments{
\item{cnv_file}{file path to the CNV file to be read}
}
\description{
Wrapper function for oce functions read.ctd, ctdTrim, and @ctdDecimate.
Extracts and adds metadata from header that is unique to SEA headers.
}
