% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{read_datasheet}
\alias{read_datasheet}
\title{Generic reader for SEA excel data sheets}
\usage{
read_datasheet(filein, sheet = 1, n_max = 100,
  range = readxl::cell_cols(1:100), ...)
}
\arguments{
\item{filein}{xls file input}
}
\description{
Produces a standardized raw output of a read excel file. It wont standardize column names, but it will ensure that data blocks are good.
Gets called from datasheet parsers (read_hourly, read_neuston, etc.)
Shouldn't ever need to call this directly, but could do to quick-check data
}
\examples{
read_datasheet()
}
