% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{read_elg}
\alias{read_elg}
\title{Read in SEA data an ELG event file and return a well formatted output}
\usage{
read_elg(filein, forceGPS = NULL, preCheck = T, skip = 0)
}
\arguments{
\item{filein}{.elg file to be read in}

\item{clean}{logical used to determine whether certain fields are
despiked and times when system is backflooded can be removed.}
}
\description{
SEA event files contain output from a number of instruments
including GPS, flow-through, chirp, etc.
}
\examples{
read_elg()

}
