% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{read_hourly}
\alias{read_hourly}
\title{Read in hourly datasheet}
\usage{
read_hourly(filein)
}
\arguments{
\item{filein}{the hourly datasheet filepath}
}
\description{
Function produces well formatted data from reading in a SEA hourlywork excel datasheet.
This is houly resolution cruise track and sensor data.
}
