% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{read_hydrocast}
\alias{read_hydrocast}
\title{Read a hydrocast datasheet}
\usage{
read_hydrocast(filein)
}
\arguments{
\item{filein}{the hydrowork datasheet}
}
\description{
Reads in formatted data from a SEA hydrowork excel datasheet. These datasheets record
bottle and auxilary data from deployment of a CTD carousel.
}
