% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{read_surfsamp}
\alias{read_surfsamp}
\title{Read Surface Sample Datasheet}
\usage{
read_surfsamp(filein)
}
\arguments{
\item{filein}{}
}
\description{
Function produces well formatted data from reading in a SEA surfsamp excel datasheet.
These datasheets contain extracted chemical and auxilary data form surface samples
collected on the SEA vessels.
}
