% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readbioll}
\alias{readbioll}
\title{Extract just the lat and lon of the neuston tows stored in a data frame}
\usage{
readbioll(df, X = NULL)
}
\arguments{
\item{df}{data frame containing Neuston tow data}

\item{X}{previousy extracted locations from another data source (e.g. CTDs)}
}
\description{
Mostly for use in plotting locations of neustons on cruise map
}
\examples{
readbioll()
}
