% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{readCTDsll}
\alias{readCTDsll}
\title{Extract just the lat and lon of the CTD profiles in a CTD list object}
\usage{
readCTDsll(CTDs, X = NULL)
}
\arguments{
\item{CTDs}{CTD list object}

\item{X}{previousy extracted locations from another data source (e.g. Neustron tows)}
}
\description{
Mostly for use in plotting locations of CTDs on cruise map
}
\examples{
readCTDsll()
}
