% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readLatLon}
\alias{readLatLon}
\title{Read lat and Lon from header of CTD CNV file}
\usage{
readLatLon(filein)
}
\arguments{
\item{filein}{.cnv file to be read from}
}
\description{
Function searches reg exp combinations to find the lat and lon of the CTD
}
\examples{
readLatLon()

}
