% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readRBRtow}
\alias{readRBRtow}
\title{Read data from an RBR used as a Tow-Yo}
\usage{
readRBRtow(filein, elg = NULL)
}
\arguments{
\item{filein}{.dat RBR file to be read from}

\item{elg}{option .elg file to be read in for locations.}
}
\description{
Function reads in data, cuts it up into casts and outputs.
If ELG file is also provided, it can assign these casts
to GPS locations.
}
\examples{
readLatLon()

}
