% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_adcp.R
\name{readSEAadcp}
\alias{readSEAadcp}
\title{Read in adcp data files in the format exported before conversion to ODV}
\usage{
readSEAadcp(filein)
}
\arguments{
\item{filein}{.txt file to read from}
}
\description{
This needs to be updated to instread read directly from the .ENS files to skip steps in on-board processing
}
\examples{
readSEAadcp()

}
