% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readSEAcsv}
\alias{readSEAcsv}
\title{Read in SEA data from a CSV file}
\usage{
readSEAcsv(filein)
}
\arguments{
\item{filein}{.csv file to read from}
}
\description{
Function used commonly as quick way of taking SEA data previously
converted from xls to csv using readSEAxls()
}
\details{
Notes: Is called automatically from readSEAxls() if the csv file is located in the same
directory as the xls file, unless rplcsv is set to T.
}
\examples{
readSEAcsv()

}
