% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readSEActd}
\alias{readSEActd}
\title{A function that reads in all .cnv files from a directory}
\usage{
readSEActd(foldin, newFL = T, CTDflag = NULL, plotFL = F,
  plotfold = "~/Desktop/")
}
\arguments{
\item{foldin}{input folder path}

\item{newFL}{Prescribe whether new .cnv files have been added or if you can just load the previously sved R file.}

\item{CTDflag}{List of indices that are to be removed from the list}

\item{plotFL}{T/F should the CTDs be plotted to folder?}

\item{plotfold}{The folder where the CTDs are to be plotted}
}
\description{
A function that reads in all .cnv files from a directory
}
\examples{
readSEActd()
}
