% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSEAdata.R
\name{readSEAelg}
\alias{readSEAelg}
\title{Read in SEA data an ELG event file}
\usage{
readSEAelg(filein, clean = T)
}
\arguments{
\item{filein}{.elg file to be read in}

\item{clean}{logical used to determine whether certain fields are
despiked and times when system is backflooded can be removed.}
}
\description{
SEA event files contain output from a number of instruments
including GPS, flow-through, chirp, etc.
}
\examples{
readSEAelg()

}
