% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{readSEAxls}
\alias{readSEAxls}
\title{Read in SEA data from an Excel spreadsheet}
\usage{
readSEAxls(filein, skip = 0, sheet = 1, rplcsv = FALSE)
}
\arguments{
\item{filein}{.xls, .xlsm, .xlsx file to read from}

\item{skip}{Number of rows to skip at the top of the xls file before reading}

\item{sheet}{Number of the sheet in the file to be read}

\item{rplcsv}{Function will create a csv file in the same directory as the
xls file for quicker reading next time. rplcsv will override whether this function
will overwrite that or look for it and read from it.}
}
\description{
Function used commonly as much data entry and organization is acheived in this format
}
\details{
Notes: Requires readxl package. readxl is slower than read.table so the function creates
a .csv file of the output for quicker reading next time. Can be overridded using rplcsv.
}
\examples{
# Read in an example neuston datasheet
df <- readSEAxls(system.file('extdata','S275_Neuston.xlsm',package='sea'))

}
