% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAplots.R
\name{reg2latlon}
\alias{reg2latlon}
\title{Set or get region data for plotting}
\usage{
reg2latlon(reg = "")
}
\arguments{
\item{reg}{Name of known region or vector of length 4 a blank charactor indicating that function should us the range of lat and lon as bounds.}
}
\description{
Function returns lat and lon limits for the plotting tools depending either
on the full range of lat and lon, or by prescribing a known region.
}
\details{
Local regions currently installed are: 'Suva', 'Nukualofa', 'lau', 'vavau'.

If prescribing a vector of length 4, this should be c(lonmin,lonmax,latmin,latmax)
}
\examples{
reg2latlon()
}
