% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rep_each}
\alias{rep_each}
\title{Repeat each element of a vector a set amount of times}
\usage{
rep_each(x, times)
}
\arguments{
\item{times}{}
}
\description{
Repeat each element of a vector a set amount of times
}
