% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{select_data}
\alias{select_data}
\title{Select a particular data source from standard data frame}
\usage{
select_data(df, data_source = "elg")
}
\arguments{
\item{data_source}{}
}
\description{
Select a particular data source from standard data frame
}
