% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{set_ll_lim}
\alias{set_ll_lim}
\title{Create good lon/lat limtis from vector}
\usage{
set_ll_lim(ll, factor = 0.15)
}
\arguments{
\item{ll}{lon or lat vector}

\item{fact}{fraction more than range}
}
\description{
Create good lon/lat limtis from vector
}
