% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{table100Count}
\alias{table100Count}
\title{Creates 100 count data summary table}
\usage{
table100Count(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates table in two parts due to length of dataset
}
\examples{
table100Count()
}
