% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{tableCTD}
\alias{tableCTD}
\title{Creates CTD summary table}
\usage{
tableCTD(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates CTD summary table
}
\examples{
tableCTD()
}
