% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{tableHydro}
\alias{tableHydro}
\title{Creates Hydrocast table}
\usage{
tableHydro(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates Hydrocast table
}
\examples{
tableHydro()
}
