% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{tableNeuston}
\alias{tableNeuston}
\title{Creates Neuston tow summary table}
\usage{
tableNeuston(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates two tables - one for overview and the second for nekton data
}
\examples{
tableNeuston)
}
