% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{tableSS}
\alias{tableSS}
\title{Creates surface station summary table}
\usage{
tableSS(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates surface station summary table
}
\examples{
tableSS()
}
