% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAtables.R
\name{tableStatSum}
\alias{tableStatSum}
\title{Creates station summary table}
\usage{
tableStatSum(filename, saveLoc = "~/Desktop")
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates station summary table
}
\examples{
tableStatSum()
}
