\name{BCFexdata}
\alias{BCFexdata}
\alias{BCF-B12}
\alias{BCF-B22}
\title{
 BCF population dataset    
}
\description{
 The phenotype of BCF population . 
}
\usage{
  data(BCFexdata)
}
\details{
 Dataset input of BCFFun function.
}
\author{
 Wang Jing-Tian, Zhang Ya-Wen, and Zhang Yuan-Ming \cr
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
