\name{G4F3exdata}
\alias{G4F3exdata}
\alias{G4F3-P1}
\alias{G4F3-P2}
\alias{G4F3-F1}
\alias{G4F3-F23}
\title{
 G4F3 population dataset    
}
\description{
 The phenotype of G4F3 population . 
}
\usage{
  data(G4F3exdata)
}
\details{
 Dataset input of G4F3Fun function.
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
