% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_layers.R
\name{make_base_map}
\alias{make_base_map}
\title{Make a base map object}
\usage{
make_base_map(df = NULL, lonlim = NULL, latlim = NULL,
  data_source = "hourly", bathy = F, bathy_legend = F, grid = T,
  high_res = F, title = "", factor = 0.15, buffer = 5)
}
\arguments{
\item{latlim}{}
}
\description{
Make a base map object
}
