% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{plot_section}
\alias{plot_section}
\title{Plot a section}
\usage{
plot_section(X, var_breaks = NULL, dist_vec = NULL, ylim = c(0, 600))
}
\arguments{
\item{var_breaks}{breaks for plotting}

\item{dist_vec}{alternative distance vector for plotting}

\item{ylim}{depth range to plot}

\item{sec}{section object made using make_section (or oce functions)}

\item{var}{variable name to plot}
}
\description{
Plot a section
}
