% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{plot_section_map}
\alias{plot_section_map}
\title{Plot a map of ctd section locations}
\usage{
plot_section_map(X, factor = 0.15)
}
\arguments{
\item{factor}{scale factor passed to set_ll_lim for setting range of plot}

\item{sec}{section object to plot}

\item{labels}{logical as to whether station IDs should be printed}

\item{...}{additional arguments to pass to make_section_map}
}
\description{
Plot a map of ctd section locations
}
