% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_surf}
\alias{plot_surf}
\title{Plot surface station data}
\usage{
plot_surf(df, var = "no3", ran_val = NULL, ran_qua = c(0, 1),
  base_map = NULL, title = var, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot surface station data
}
