% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_vars.R
\name{plot_wind}
\alias{plot_wind}
\title{Plot surface from an ELG or Hourly file}
\usage{
plot_wind(df, data_source = "elg", field = "wind", step = 60, scale = 1,
  ...)
}
\arguments{
\item{df}{}
}
\description{
Plot surface from an ELG or Hourly file
}
