% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctd.R
\name{prep_section_hydro}
\alias{prep_section_hydro}
\title{Prepare bottle-derived data to plot}
\usage{
prep_section_hydro(df, var = "chla", select = NULL, xo = NULL,
  yo = NULL, along_path = T, dist_vec = NULL)
}
\arguments{
\item{df}{data frame read in using read_hydrocast}

\item{var}{}
}
\description{
Prepare bottle-derived data to plot
}
