% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{readSEAadcp_all}
\alias{readSEAadcp_all}
\title{As readSEAadcp, this needs to be updated to instread read directly from the .ENS files to skip steps in on-board processing}
\usage{
readSEAadcp_all(foldin)
}
\arguments{
\item{foldin}{folder containing all the .txt file to read from}
}
\description{
As readSEAadcp, this needs to be updated to instread read directly from the .ENS files to skip steps in on-board processing
}
\examples{
readSEAadcp_all()

}
